nrf52: Add support for retrieving u-blox NINA-B1 public address
u-blox NINA-B1 is an nrf52 based radio module with a public address
stored in UICR registers.  It differs from the existing code in that
it is stored in Big Endian and marked with the upper 16-bits of
CUSTOMER[1] all set to 1.

This patch checks for this and converts and assembles the Public
MAC accordingly without changing the previous behaviour for other
NRF52 modules that mark the upper 16-bits with all zeros.

Tested on the u-blox NINA-B1 device integrated on the SolidRun
i.MX8M based SOM.

Signed-off-by: Jon Nettleton <jon@solid-run.com>